Added new conflict exception for entity duplicates & reorginizing project
