+++ src/main/java/com/revealprecision/revealserver/api/v1/dto/response/GeographicLevelResponse.java
@@ -1,4 +1,4 @@
-package com.revealprecision.revealserver.api.dto.response;
+package com.revealprecision.revealserver.api.v1.dto.response;
 
 import java.util.UUID;
 import lombok.AllArgsConstructor;
