+++ src/main/java/com/revealprecision/revealserver/exceptions/ConflictException.java
@@ -1,0 +1,8 @@
+package com.revealprecision.revealserver.exceptions;
+
+public class ConflictException extends RuntimeException {
+
+  public ConflictException(String message) {
+    super(message);
+  }
+}
