+++ src/main/java/com/revealprecision/revealserver/exceptions/constant/Error.java
@@ -7,4 +7,5 @@ import lombok.NoArgsConstructor;
 public class Error {
 
   public static final String RESOURCE_NOT_FOUND = "Resource '%s' with '%s' not found";
+  public static final String NON_UNIQUE = "%s with value '%s' already exists";
 }
