+++ src/main/java/com/revealprecision/revealserver/persistence/repository/GeographicLevelRepository.java
@@ -1,12 +1,13 @@
 package com.revealprecision.revealserver.persistence.repository;
 
 import com.revealprecision.revealserver.persistence.domain.GeographicLevel;
+import java.util.Optional;
+import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
-public interface GeographicLevelRepository extends JpaRepository<GeographicLevel,UUID>{
-    GeographicLevel findByName(String name);
+public interface GeographicLevelRepository extends JpaRepository<GeographicLevel, UUID> {
+
+  Optional<GeographicLevel> findByName(String name);
 }
