+++ src/main/resources/application.properties
@@ -1,62 +1,51 @@
 # -----------------------------------------------------------------------------------------------------------------------
 # general configs
 # -----------------------------------------------------------------------------------------------------------------------
-server.port = 8080
-server.max-http-header-size = 45000
-
-spring.application.name = revealserver
-
+server.port=8080
+server.max-http-header-size=45000
+spring.application.name=revealserver
 # -----------------------------------------------------------------------------------------------------------------------
 # database configs
 # -----------------------------------------------------------------------------------------------------------------------
-spring.datasource.driver-class-name = org.postgresql.Driver
-spring.datasource.url = jdbc:postgresql://localhost:5500/revealserver
-spring.datasource.username = postgres
-spring.datasource.password = NjNjYWU3YjY5MmZhNGMwOWMxOTQ2YWVm
-
-spring.datasource.hikari.connection-timeout = 10000
-spring.datasource.hikari.driver-class-name = org.postgresql.Driver
-spring.datasource.hikari.maximum-pool-size = 10
-
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-spring.jackson.time-zone = CAT
-
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.datasource.url=jdbc:postgresql://10.0.0.19:5432/revealserver
+spring.datasource.username=postgres
+spring.datasource.password=NjNjYWU3YjY5MmZhNGMwOWMxOTQ2YWVm
+spring.datasource.hikari.connection-timeout=10000
+spring.datasource.hikari.driver-class-name=org.postgresql.Driver
+spring.datasource.hikari.maximum-pool-size=10
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jackson.time-zone=CAT
 spring.jpa.properties.org.hibernate.envers.store_data_at_delete=true
-
-hibernate.types.print.banner = false
-
+hibernate.types.print.banner=false
 # -----------------------------------------------------------------------------------------------------------------------
 # kafka configs
 # -----------------------------------------------------------------------------------------------------------------------
-spring.kafka.consumer.bootstrap-servers = kafka-01:29092,kafka-02:29093,kafka-03:29094
-spring.kafka.consumer.group-id = reveal_server_group
-spring.kafka.consumer.auto-offset-reset = latest
-spring.kafka.consumer.key-deserializer = org.apache.kafka.common.serialization.StringDeserializer
-spring.kafka.consumer.value-deserializer = org.apache.kafka.common.serialization.StringDeserializer
-
-spring.kafka.producer.bootstrap-servers = kafka-01:29092,kafka-02:29093,kafka-03:29094
-spring.kafka.producer.key-serializer = org.apache.kafka.common.serialization.StringSerializer
-spring.kafka.producer.value-serializer = org.apache.kafka.common.serialization.StringSerializer
-
+spring.kafka.consumer.bootstrap-servers=kafka-01:29092,kafka-02:29093,kafka-03:29094
+spring.kafka.consumer.group-id=reveal_server_group
+spring.kafka.consumer.auto-offset-reset=latest
+spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+spring.kafka.producer.bootstrap-servers=kafka-01:29092,kafka-02:29093,kafka-03:29094
+spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
+spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.StringSerializer
 # -----------------------------------------------------------------------------------------------------------------------
 # tracing configs
 # -----------------------------------------------------------------------------------------------------------------------
- spring.sleuth.sampler.probability=1.0
- spring.zipkin.sender.type=kafka
-
+spring.sleuth.sampler.probability=1.0
+spring.zipkin.sender.type=kafka
 # -----------------------------------------------------------------------------------------------------------------------
 # keycloak configs
 # -----------------------------------------------------------------------------------------------------------------------
-keycloak.enabled = true
-keycloak.auth-server-url = https://sso-ops.akros.online/auth
-keycloak.realm = reveal
-keycloak.resource = reveal-server
-keycloak.credentials.secret = 380cd0be-e0bb-4867-b5c6-24776d05cc1a
-keycloak.public-client = false
-keycloak.ssl-required = external
-keycloak.use-resource-role-mappings = false
-keycloak.bearer-only = true
-
+keycloak.enabled=true
+keycloak.auth-server-url=https://sso-ops.akros.online/auth
+keycloak.realm=reveal
+keycloak.resource=reveal-server
+keycloak.credentials.secret=380cd0be-e0bb-4867-b5c6-24776d05cc1a
+keycloak.public-client=false
+keycloak.ssl-required=external
+keycloak.use-resource-role-mappings=false
+keycloak.bearer-only=true
 keycloak.security-constraints[0].authRoles[0]=plan_view
 keycloak.security-constraints[0].security-collections[0].methods[0]=GET
 keycloak.security-constraints[0].security-collections[0].patterns[0]=/api/v1/plan
@@ -108,31 +97,27 @@ keycloak.security-constraints[11].authRoles[0]=geographic_level_view
 keycloak.security-constraints[11].authRoles[1]=geographic_level_delete
 keycloak.security-constraints[11].security-collections[0].methods[0]=DELETE
 keycloak.security-constraints[11].security-collections[0].patterns[0]=/api/v1/geographicLevel/{identifier}
-
 # -----------------------------------------------------------------------------------------------------------------------
 # cors configs
 # -----------------------------------------------------------------------------------------------------------------------
-
 # -----------------------------------------------------------------------------------------------------------------------
 # actuator configs
 # -----------------------------------------------------------------------------------------------------------------------
 management.endpoints.health.sensitive=false
 management.endpoint.health.show-details=always
-
 # -----------------------------------------------------------------------------------------------------------------------
 # logging configs
 # -----------------------------------------------------------------------------------------------------------------------
-logging.level.root = INFO
-logging.level.org.springframework = INFO
-logging.level.org.springframework.web.servlet.mvc.method.annotation = TRACE
-logging.level.org.apache.kafka.clients = WARN
-logging.level.org.keycloak = TRACE
-
+logging.level.root=INFO
+logging.level.org.springframework=INFO
+logging.level.org.springframework.web.servlet.mvc.method.annotation=TRACE
+logging.level.org.apache.kafka.clients=WARN
+logging.level.org.keycloak=TRACE
 # -----------------------------------------------------------------------------------------------------------------------
 # SpringDoc: Swagger
 # -----------------------------------------------------------------------------------------------------------------------
-springdoc.api-docs.enabled = true
-springdoc.api-docs.path = /api-docs
-springdoc.version = 3.0.0
-springdoc.swagger-ui.enabled = false
-spring.main.allow-bean-definition-overriding = false
+springdoc.api-docs.enabled=true
+springdoc.api-docs.path=/api-docs
+springdoc.version=3.0.0
+springdoc.swagger-ui.enabled=false
+spring.main.allow-bean-definition-overriding=false
